+++ src/lib/motion-start/render/dom/Motion-Proxy.svelte
@@ -1,7 +1,6 @@
 <!-- <svelte:options runes={true}/> -->
 <script lang="ts">
     import type { SvelteHTMLElements } from "svelte/elements";
-    import type { Snippet } from "svelte";
     import type { MotionProps } from "../../motion";
     import Motion from "../../motion/Motion.svelte";
 
@@ -13,10 +12,10 @@
 <Motion {...$$restProps} let:props let:motion>
     <svelte:element
         this={___tag}
+        {...props}
         bind:this={el}
         class={props.class}
         use:motion
-        {...props}
     >
         <slot />
     </svelte:element>