+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # motion-start
 
+## 0.1.3
+
+### Patch Changes
+
+- 61bcd64: fix export path for svelte comp
+
+## 0.1.2
+
+### Patch Changes
+
+- 25f1b6e: export MotionSSR
+
+## 0.1.1
+
+### Patch Changes
+
+- 98cfaba: chore: 🤖 release
+- 4f1b2e7: enable layout animation feature
+- 9f5a69c: fix layout animations
+
 ## 0.1.0
 
 ### Minor Changes